add analyzer for number supertype binary operations
