+++ runtime/sema/check_binary_expression.go
@@ -26,6 +26,14 @@ import (
 
 func (checker *Checker) VisitBinaryExpression(expression *ast.BinaryExpression) ast.Repr {
 
+	var leftType, rightType, resultType Type
+	defer func() {
+		elaboration := checker.Elaboration
+		elaboration.BinaryExpressionLeftTypes[expression] = leftType
+		elaboration.BinaryExpressionRightTypes[expression] = rightType
+		elaboration.BinaryExpressionResultTypes[expression] = resultType
+	}()
+
 	// The left-hand side is always evaluated.
 	// However, the right-hand side might not necessarily be evaluated,
 	// e.g. in boolean logic or in nil-coalescing
@@ -55,7 +63,7 @@ func (checker *Checker) VisitBinaryExpression(expression *ast.BinaryExpression) 
 	// Visit the expression, with contextually expected type. Use the expected type
 	// only for inferring wherever possible, but do not check for compatibility.
 	// Compatibility is checked separately for each operand kind.
-	leftType := checker.VisitExpressionWithForceType(expression.Left, expectedType, false)
+	leftType = checker.VisitExpressionWithForceType(expression.Left, expectedType, false)
 
 	leftIsInvalid := leftType.IsInvalidType()
 
@@ -111,7 +119,7 @@ func (checker *Checker) VisitBinaryExpression(expression *ast.BinaryExpression) 
 			expectedType = leftType
 		}
 
-		rightType := checker.VisitExpressionWithForceType(expression.Right, expectedType, false)
+		rightType = checker.VisitExpressionWithForceType(expression.Right, expectedType, false)
 
 		rightIsInvalid := rightType.IsInvalidType()
 
@@ -122,18 +130,20 @@ func (checker *Checker) VisitBinaryExpression(expression *ast.BinaryExpression) 
 			BinaryOperationKindNonEqualityComparison,
 			BinaryOperationKindBitwise:
 
-			return checker.checkBinaryExpressionArithmeticOrNonEqualityComparisonOrBitwise(
+			resultType = checker.checkBinaryExpressionArithmeticOrNonEqualityComparisonOrBitwise(
 				expression, operation, operationKind,
 				leftType, rightType,
 				leftIsInvalid, rightIsInvalid, anyInvalid,
 			)
+			return resultType
 
 		case BinaryOperationKindEquality:
-			return checker.checkBinaryExpressionEquality(
+			resultType = checker.checkBinaryExpressionEquality(
 				expression, operation,
 				leftType, rightType,
 				anyInvalid,
 			)
+			return resultType
 
 		default:
 			return unsupportedOperation()
@@ -146,7 +156,7 @@ func (checker *Checker) VisitBinaryExpression(expression *ast.BinaryExpression) 
 		// That means that resource invalidation and returns
 		// are not definite, but only potential.
 
-		rightType := checker.checkPotentiallyUnevaluated(func() Type {
+		rightType = checker.checkPotentiallyUnevaluated(func() Type {
 			var expectedType Type
 			if !leftIsInvalid {
 				if optionalLeftType, ok := leftType.(*OptionalType); ok {
@@ -162,22 +172,19 @@ func (checker *Checker) VisitBinaryExpression(expression *ast.BinaryExpression) 
 
 		switch operationKind {
 		case BinaryOperationKindBooleanLogic:
-			return checker.checkBinaryExpressionBooleanLogic(
+			resultType = checker.checkBinaryExpressionBooleanLogic(
 				expression, operation,
 				leftType, rightType,
 				leftIsInvalid, rightIsInvalid, anyInvalid,
 			)
+			return resultType
 
 		case BinaryOperationKindNilCoalescing:
-			resultType := checker.checkBinaryExpressionNilCoalescing(
+			resultType = checker.checkBinaryExpressionNilCoalescing(
 				expression, operation,
 				leftType, rightType,
 				leftIsInvalid, rightIsInvalid,
 			)
-
-			checker.Elaboration.BinaryExpressionResultTypes[expression] = resultType
-			checker.Elaboration.BinaryExpressionRightTypes[expression] = rightType
-
 			return resultType
 
 		default:
